requirements.txt
pytesseract==0.3.13
python-telegram-bot==21.4
pillow==10.4.0
Telethon==1.36.0